requirements.txt
gspread~=6.2.1
aiogram~=3.20.0.post0
python-dotenv~=1.1.0
pandas~=2.2.3
yfinance~=0.2.61
curl_cffi~=0.11.1
aiohttp~=3.11.18
google-genai
sqlalchemy
asyncpg
asyncio
redis